basic plumbing for network-specific explore pages
