+++ src/routes/app/(app)/components/load-filecoin-explore-page-data.ts
@@ -1,0 +1,3 @@
+export default async function loadFilecoinExporePageData() {
+  return {};
+}
